Suppress ant warning if missing extension lib dir
Many of the extensions do not include external dependencies, but for
consistency in the build file all extensions attempt to copy libs if
they exist. The copy tasks were already to configured to not fail if
there was an error copying the files (because the source dir does not
exist in this case.) This change also sets the quiet option so that a
warning is not produced in the build output on a copy error.

Signed-off-by: Tony Germano <tony@germano.name>