Add more module options to default vmoptions file (#194)
If you click the Get Operations button on a Web Service Sender with a bad URL, the server will throw an exception. The chained exceptions get serialized to send back to the client, and that is where these --add-opens are required. The exceptions include both a java.util.concurrent.ExecutionException and a java.io.IOException.

Signed-off-by: Kaur Palang <kaur.palang@brightcodecompany.com>