Update XMLReader plugin ContentHandler calls
According to the javadoc for ContentHandler.startElement, if there are
no attributes, then an empty Attributes object should be passed rather
than null.

For both startElement and endElement if qualified names are not
available then the qname parameter should be an empty string. It was
already being used this way in most instances, but a few were passing
null instead.

Signed-off-by: Tony Germano <tony@germano.name>